requirements.txt
pip~=25.0
wheel~=0.41.2
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.10.6
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic-settings~=2.7.1
cryptography~=44.0.1